+++ src/Input/Input.stories.tsx
@@ -1,6 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react'
 
 import Input from './Input'
+import SearchIcon from '../../icons/system/search.svg'
 
 export default {
   title: 'Atoms/Input',
@@ -12,7 +13,7 @@ export default {
       },
     },
   },
-  args: { inputId: 'test-input', label: 'Email', required: true },
+  args: { inputId: 'test-input', label: 'Email', required: true, icon: <SearchIcon/> },
   argTypes: {
     label: {
       type: 'string',