+++ CHANGELOG.md
@@ -13,6 +13,15 @@ Types of changes are:
 6. Security - For security features
 
 ## Unreleased
+### Added
+- New system icons (plus, search, trash, edit, data, close)
+- New mono-color icon (check)
+- `_custom.scss` css theme file
+
+### Changed
+- Added icons props to `Toast` component
+- Added icons props to `Button` component
+- Added icons props to `Input` component
 
 ## 1.0.27 - 2025-03-12
 ### Added